crawler/4traffic_sensors/requirements.txt
paho_mqtt
psycopg2_binary
